backend/cartpe/requirements.txt
amqp==5.2.0
asgiref==3.8.1
astroid==3.2.3
async-timeout==4.0.3
billiard==4.2.0
black==24.4.2
celery==5.4.0
certifi==2024.7.4
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
construct==2.10.70
coverage==7.6.0
dill==0.3.8
distlib==0.3.8
Django==5.0.7
django-cors-headers==4.4.0
django-filter==24.2
django-haystack==3.3.0
django-js-asset==2.2.0
django-mptt==0.16.0
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-camel-case==1.4.2
djangorestframework-recursive==0.1.2
djangorestframework-simplejwt==5.3.1
filelock==3.15.4
identify==2.6.0
idna==3.7
isort==5.13.2
kombu==5.3.7
lazy-object-proxy==1.10.0
Markdown==3.6
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.1
pathspec==0.12.1
pefile==2023.2.7
pillow==10.4.0
pip-review==1.3.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
PyJWT==2.8.0
pylint==3.2.5
pysolr==3.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-ptrace==0.9.9
pytz==2024.1
PyYAML==6.0.1
razorpay==1.4.2
redis==5.0.7
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
tomlkit==0.13.0
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
virtualenv==20.26.3
wcwidth==0.2.13
wrapt==1.16.0
